GTB-77 [refactor] 도메인 수정 및 스타일 통일

+++ src/main/java/site/gachontable/gachontablebe/domain/admin/presentation/dto/request/ExitUserRequest.java
@@ -1,4 +1,4 @@
 package site.gachontable.gachontablebe.domain.admin.presentation.dto.request;
 
-public record ExitUserRequest(Integer seatingId) {
+public record ExitUserRequest(Long seatingId) {
 }
